# backend/swagger.json
{
  "openapi": "3.0.1",
  "servers": [
    {
      "url": "https://ysp-api.vs-blog.tech",
      "description": "Local Server"
    }
  ],
  "info": {
    "version": "1.0.0",
    "title": "ysp",
    "description": "Nest TypeScript starter repository",
    "license": {
      "name": "MIT"
    }
  },
  "paths": {
...
    "/auth/signup": {
      "post": {
        "tags": [
          "Auth"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResult"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AuthBody": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1
          },
          "password": {
            "type": "string",
            "maxLength": 50,
            "minLength": 8
          }
        },
        "nullable": false,
        "required": [
          "username",
          "password"
        ]
      },
      "LoginResult": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "access_token_ttl": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          },
          "refresh_token_ttl": {
            "type": "string"
          }
        },
        "nullable": false,
        "required": [
          "access_token",
          "access_token_ttl",
          "refresh_token",
          "refresh_token_ttl"
        ]
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}